Fix yesterday filter  - Uses [yesterday, yesterday] as the period
