+++ src/pages/Notifications/EventLog/EventLogPage.tsx
@@ -1,7 +1,5 @@
 import { Text, TextContent } from '@patternfly/react-core';
 import { Main, PageHeader, PageHeaderTitle } from '@redhat-cloud-services/frontend-components';
-import { Filter, Operator, toUtc } from '@redhat-cloud-services/insights-common-typescript';
-import { format, sub, toDate } from 'date-fns';
 import * as React from 'react';
 
 import { EventLogDateFilterValue } from '../../../components/Notifications/EventLog/EventLogDateFilter';
@@ -15,9 +13,9 @@ import { useGetApplications } from '../../../services/Notifications/GetApplicati
 import { useGetBundles } from '../../../services/Notifications/GetBundles';
 import { EventPeriod } from '../../../types/Event';
 import { useEventLogFilter } from './useEventLogFilter';
+import { useFilterBuilder } from './useFilterBuilder';
 
 const RETENTION_DAYS = 14;
-const DATE_FORMAT = 'yyyy-MM-dd';
 
 export const EventLogPage: React.FunctionComponent = () => {
 
@@ -47,59 +45,7 @@ export const EventLogPage: React.FunctionComponent = () => {
 
     const [ period, setPeriod ] = React.useState<EventPeriod>([ undefined, undefined ]);
 
-    const filterBuilder = React.useCallback((filters?: EventLogFilters) => {
-        const filter = new Filter();
-        if (filters?.bundle) {
-            const selectedBundleNames = filters?.bundle;
-            const selectedBundles = bundles.filter(b => selectedBundleNames.includes(b.name)).map(b => b.id);
-            filter.and('bundleIds', Operator.EQUAL, selectedBundles);
-        }
-
-        if (filters?.application) {
-            const selectedAppNames = filters?.application;
-            const selectedApps = applications.filter(a => selectedAppNames.includes(a.name)).map(a => a.id);
-            filter.and('appIds', Operator.EQUAL, selectedApps);
-        }
-
-        if (filters?.event) {
-            filter.and('event', Operator.EQUAL, filters.event);
-        }
-
-        let filterPeriod = [ undefined, undefined ] as [ Date | undefined, Date | undefined ];
-        const today = toUtc(new Date());
-
-        switch (dateFilter) {
-            case EventLogDateFilterValue.LAST_14:
-                filterPeriod = [ sub(toDate(today), {
-                    days: 14
-                }), today ];
-                break;
-            case EventLogDateFilterValue.LAST_7:
-                filterPeriod = [ sub(toDate(today), {
-                    days: 7
-                }), today ];
-                break;
-            case EventLogDateFilterValue.TODAY:
-                filterPeriod = [ today, today ];
-                break;
-            case EventLogDateFilterValue.YESTERDAY:
-                filterPeriod = [ sub(toDate(today), {
-                    days: 1
-                }), today ];
-                break;
-            case EventLogDateFilterValue.CUSTOM:
-                filterPeriod = period;
-                break;
-        }
-
-        if (filterPeriod[0] && filterPeriod[1]) {
-            filter.and('start', Operator.EQUAL, format(filterPeriod[0], DATE_FORMAT));
-            filter.and('end', Operator.EQUAL, format(filterPeriod[1], DATE_FORMAT));
-        }
-
-        return filter;
-    }, [ bundles, applications, dateFilter, period ]);
-
+    const filterBuilder = useFilterBuilder(bundles, applications, dateFilter, period);
     const eventsPage = usePage<EventLogFilters>(10, filterBuilder, eventLogFilters.filters);
 
     const eventsQuery = useGetEvents(eventsPage.page);
